Adding the necesarry files to build the project
Adding the necesarry files to build the project. In addition I added
some configs that I know is necesarry to run the SPI to communicate with
the external memory. After this commit one can easily build the project
using west (if west is not found, you might need to set an envirenment
variable names ZEPHYR_BASE to the path to the base of the zephyr
directory you have installed on your device). Flashing guidelines will
come soon.

Files added:
- .gitignore: Ignore build directory
- CMakeLists.txt: Add to find zephyr and source files
- boards/nrf52833dk_nrf52833.conf: Board-specific configurations
- boards/nrf52833dk_nrf52833.overlay: Overwriting default hardware
settings
- prj.conf: Project-specific configurations
- src/main.c: Add simple main

CT-169