Basic implementation of scanning and advertising
This is a basic implementation of the scan, advertise, ble and
connection module. This will serve as a base for further development.
There is a lot of missing functionality that will be implemented later.
Having this simple implementation of the modules will make it easier
to develop code for the different modules. It seems that there is no
need for a GAP module, but a connection module is necessary to handle
the connections. This will be added to the class diagram.

Files added:
- src/ble/advertise.c: Basic implementation of the advertise module
- src/ble/advertise.h: Basic implementation of the advertise module
- src/ble/ble.c: Basic implementation of the ble module
- src/ble/ble.h: Basic implementation of the ble module
- src/ble/connection.c: Basic implementation of the connection module
- src/ble/connection.h: Basic implementation of the connection module
- src/ble/scan.c: Basic implementation of the scan module
- src/ble/scan.h: Basic implementation of the scan module

Files modified:
- CMakeLists.txt: Add the source files
- prj.conf: Add configurations required by the modules
- src/main.c: Add ble init and includes

CT-171