Implement gaens crypto functions. Rename module
Finish implementation of all the cryptography functions from the
GAENS specification. Has not been tested yet, but more framework is
needed before testing is relevant so this PR is added in order to be
able to start work on that soon. There will have to be another module
that keeps track of when to use all of these functions etc., and that
should probably be a separate module as well, with easy function calls
that can be run from the bluetooth modules which uses this.

Renamed the module to crypto, as there will be more modules inside the
gaens folder.

Files added:
- src/gaens/crypto.c: Implement the gaens crypto functions
- src/gaens/crypto.h: Document with doxygen the functions

Files modified:
- .gitignore: Add .vscode folder
- CMakeLists.txt: Add crypto module build target
- prj.conf: Add neccessary build libraries

Files deleted:
- src/gaens/gaens.c: Renamed gaens to crypto
- src/gaens/gaens.h: Renamed gaens to crypto

CT-175