Implement scan filtering
This is a basic implementation of scan filtering. It will have to be
changed later when we add more services and finish more of the other
modules. This implementation filter out GAENS packets (and WENS packets
when we change to the correct UUID). I have added notes in the file to
make it easier for future changes.

Files modified:
- src/ble/scan.c: Implement scan filtering

CT-171