Fix GAENS and advertising mistake
This commit fixes a mistake in gaens.c where I called a function
that did not need to be called. I also fix a mistake in the GAENS
advertising packets. I forgot that the UUID of the service data
has to be included together with the data.

Files modified:
- src/ble/advertise.c: Small fix
- src/gaens/gaens.c: Small fix

CT-166